Move DiagnosticStateNonHydro to common/states
The driver allocates it, the initial condition fills part of it, and both
solve_nonhydro and velocity_advection mutate it: it is an interface between the
driver and the granule, like the prognostic and diagnostic states that already
live in common/states.

It is moved, not re-exported: there is one home for it.
