+++ model/atmosphere/dycore/tests/dycore/integration_tests/test_benchmark_solve_nonhydro.py
@@ -31,7 +31,11 @@ from icon4py.model.common.grid import (
 )
 from icon4py.model.common.interpolation import interpolation_attributes, interpolation_factory
 from icon4py.model.common.metrics import metrics_attributes, metrics_factory
-from icon4py.model.common.states import factory, prognostic_state as prognostics
+from icon4py.model.common.states import (
+    factory,
+    nonhydro_diagnostic_state as nonhydro_states,
+    prognostic_state as prognostics,
+)
 from icon4py.model.common.utils import data_allocation as data_alloc
 from icon4py.model.testing.fixtures.benchmark import (
     geometry_field_source,
@@ -243,7 +247,7 @@ def test_benchmark_solve_nonhydro(  # noqa: PLR0917 [too-many-positional-argumen
         ),
     )
 
-    diagnostic_state_nh = dycore_states.DiagnosticStateNonHydro(
+    diagnostic_state_nh = nonhydro_states.DiagnosticStateNonHydro(
         max_vertical_cfl=data_alloc.scalar_like_array(0.0, allocator),
         theta_v_at_cells_on_half_levels=data_alloc.zero_field(
             mesh, dims.CellDim, dims.KDim, extend={dims.KDim: 1}, allocator=allocator
